refactor(db): CountrySubdivisionRepository update_batch fix, code cleanup, and EntityReferenceRepository decomposition
This commit includes a fix for the `CountrySubdivisionRepositoryImpl::update_batch` function, general code cleanup, and a significant refactoring of the `EntityReferenceRepositoryImpl`.

The `CountrySubdivisionRepositoryImpl::update_batch` function was corrected to properly handle updates to all fields, not just indexed ones, resolving a bug where non-indexed field changes were ignored. This also involved cleaning up unused imports and mutable variables as suggested by the Rust compiler.

The `EntityReferenceRepositoryImpl` has been decomposed into individual, single-responsibility files within a dedicated directory (`repository/person/entity_reference_repository/`). This refactoring improves modularity, maintainability, and testability by breaking down the monolithic implementation into smaller, focused files for each database operation. Corresponding tests have been adjusted or co-located to align with the new modular structure.

Components affected:
- banking-db-postgres:
    - `CountrySubdivisionRepositoryImpl`: Fixes `update_batch` logic and general code cleanup.
    - `EntityReferenceRepositoryImpl`: Decomposition into modular files.
    - Various files: Removal of unused imports and adjustment of variable mutability.
- docs: Updates the progress tracking document to reflect these achievements.

Benefits:
- Corrected `update_batch` functionality for `CountrySubdivisionModel`.
- Improved code quality and reduced compiler warnings.
- Enhanced modularity, maintainability, and testability of `EntityReferenceRepository`.
- Reduced file sizes for better readability and LLM performance.
- Validated fixes with passing tests.
