+++ banking-db-postgres/src/postgres_repositories.rs
@@ -13,7 +13,7 @@ use crate::repository::{
     person::{
         country_repository::repo_impl::CountryRepositoryImpl,
         country_subdivision_repository::CountrySubdivisionRepositoryImpl,
-        entity_reference_repository_impl::EntityReferenceRepositoryImpl,
+        entity_reference_repository::EntityReferenceRepositoryImpl,
         locality_repository_impl::LocalityRepositoryImpl,
         location_repository_impl::LocationRepositoryImpl,
         person_repository_impl::PersonRepositoryImpl,
