+++ banking-db-postgres/src/repository/person/country_subdivision_repository/find_by_id.rs
@@ -17,7 +17,7 @@ pub async fn find_by_id(
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    
     use crate::repository::person::test_helpers::{
         create_test_country_model, create_test_country_subdivision_model,
     };
