+++ banking-db-postgres/src/repository/person/country_subdivision_repository/save.rs
@@ -99,7 +99,7 @@ pub async fn save(
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    
     use crate::repository::person::test_helpers::{
         create_test_country_model, create_test_country_subdivision_model,
     };
