--- banking-db-postgres/src/repository/person/entity_reference_repository_impl.rs
@@ -1,689 +1,0 @@
-use async_trait::async_trait;
-use banking_api::BankingResult;
-use crate::utils::{get_heapless_string, get_optional_heapless_string, TryFromRow};
-use banking_db::models::person::{
-    EntityReferenceAuditModel, EntityReferenceIdxModel, EntityReferenceIdxModelCache,
-    EntityReferenceModel,
-};
-use banking_db::repository::{EntityReferenceRepository, PersonRepository, TransactionAware};
-use banking_db::repository::person::entity_reference_repository::{EntityReferenceRepositoryError, EntityReferenceResult};
-use crate::repository::executor::Executor;
-use crate::repository::person::person_repository_impl::PersonRepositoryImpl;
-use sqlx::{postgres::PgRow, Postgres, Row};
-use std::collections::{HashMap, HashSet};
-use std::error::Error;
-use std::hash::Hasher;
-use std::sync::Arc;
-use parking_lot::RwLock;
-use tokio::sync::RwLock as TokioRwLock;
-use twox_hash::XxHash64;
-use uuid::Uuid;
-
-pub struct EntityReferenceRepositoryImpl {
-    pub(crate) executor: Executor,
-    pub(crate) entity_reference_idx_cache:
-        Arc<TokioRwLock<TransactionAwareEntityReferenceIdxModelCache>>,
-    person_repository: Arc<PersonRepositoryImpl>,
-}
-
-impl EntityReferenceRepositoryImpl {
-    pub fn new(
-        executor: Executor,
-        person_repository: Arc<PersonRepositoryImpl>,
-        entity_reference_idx_cache: Arc<RwLock<EntityReferenceIdxModelCache>>,
-    ) -> Self {
-        Self {
-            executor,
-            entity_reference_idx_cache: Arc::new(TokioRwLock::new(
-                TransactionAwareEntityReferenceIdxModelCache::new(entity_reference_idx_cache),
-            )),
-            person_repository,
-        }
-    }
-
-    pub async fn load_all_entity_reference_idx(
-        executor: &Executor,
-    ) -> Result<Vec<EntityReferenceIdxModel>, sqlx::Error> {
-        let query =
-            sqlx::query_as::<_, EntityReferenceIdxModel>("SELECT * FROM entity_reference_idx");
-        match executor {
-            Executor::Pool(pool) => query.fetch_all(&**pool).await,
-            Executor::Tx(tx) => {
-                let mut tx = tx.lock().await;
-                query.fetch_all(&mut **tx).await
-            }
-        }
-    }
-}
-
-#[async_trait]
-impl EntityReferenceRepository<Postgres> for EntityReferenceRepositoryImpl {
-    async fn save(
-        &self,
-        entity_ref: EntityReferenceModel,
-        audit_log_id: Uuid,
-    ) -> EntityReferenceResult<EntityReferenceModel> {
-        if !self
-            .person_repository
-            .exists_by_id(entity_ref.person_id)
-            .await
-            .map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?
-        {
-            return Err(EntityReferenceRepositoryError::PersonNotFound(entity_ref.person_id));
-        }
-
-        let mut hasher = XxHash64::with_seed(0);
-        let mut entity_ref_cbor = Vec::new();
-        ciborium::ser::into_writer(&entity_ref, &mut entity_ref_cbor).unwrap();
-        hasher.write(&entity_ref_cbor);
-        let new_hash = hasher.finish() as i64;
-
-        let mut ref_hasher = XxHash64::with_seed(0);
-        ref_hasher.write(entity_ref.reference_external_id.as_bytes());
-        let reference_external_id_hash = ref_hasher.finish() as i64;
-
-        let maybe_existing_idx = {
-            let cache_read_guard = self.entity_reference_idx_cache.read().await;
-            cache_read_guard.get_by_primary(&entity_ref.id)
-        };
-
-        if let Some(existing_idx) = maybe_existing_idx {
-            // UPDATE
-            if existing_idx.hash == new_hash {
-                return Ok(entity_ref); // No changes
-            }
-
-            let new_version = existing_idx.version + 1;
-
-            let audit_model = EntityReferenceAuditModel {
-                entity_reference_id: entity_ref.id,
-                version: new_version,
-                hash: new_hash,
-                person_id: entity_ref.person_id,
-                entity_role: entity_ref.entity_role,
-                reference_external_id: entity_ref.reference_external_id.clone(),
-                reference_details_l1: entity_ref.reference_details_l1.clone(),
-                reference_details_l2: entity_ref.reference_details_l2.clone(),
-                reference_details_l3: entity_ref.reference_details_l3.clone(),
-                audit_log_id,
-            };
-
-            let query1 = sqlx::query(
-                r#"
-                INSERT INTO entity_reference_audit (entity_reference_id, version, hash, person_id, entity_role, reference_external_id, reference_details_l1, reference_details_l2, reference_details_l3, audit_log_id)
-                VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
-                "#,
-            )
-            .bind(audit_model.entity_reference_id)
-            .bind(audit_model.version)
-            .bind(audit_model.hash)
-            .bind(audit_model.person_id)
-            .bind(audit_model.entity_role)
-            .bind(audit_model.reference_external_id.as_str())
-            .bind(
-                audit_model
-                    .reference_details_l1
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(
-                audit_model
-                    .reference_details_l2
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(
-                audit_model
-                    .reference_details_l3
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(audit_model.audit_log_id);
-
-            let query2 = sqlx::query(
-                r#"
-                UPDATE entity_reference SET
-                    person_id = $2, entity_role = $3::person_entity_type, reference_external_id = $4,
-                    reference_details_l1 = $5, reference_details_l2 = $6, reference_details_l3 = $7
-                WHERE id = $1
-                "#,
-            )
-            .bind(entity_ref.id)
-            .bind(entity_ref.person_id)
-            .bind(entity_ref.entity_role)
-            .bind(entity_ref.reference_external_id.as_str())
-            .bind(
-                entity_ref
-                    .reference_details_l1
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(
-                entity_ref
-                    .reference_details_l2
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(
-                entity_ref
-                    .reference_details_l3
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            );
-
-            let query3 = sqlx::query(
-                r#"
-                UPDATE entity_reference_idx SET
-                    version = $2,
-                    hash = $3
-                WHERE entity_reference_id = $1
-                "#,
-            )
-            .bind(entity_ref.id)
-            .bind(new_version)
-            .bind(new_hash);
-
-            match &self.executor {
-                Executor::Pool(pool) => {
-                    query1.execute(&**pool).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                    query2.execute(&**pool).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                    query3.execute(&**pool).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                }
-                Executor::Tx(tx) => {
-                    let mut tx = tx.lock().await;
-                    query1.execute(&mut **tx).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                    query2.execute(&mut **tx).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                    query3.execute(&mut **tx).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                }
-            }
-
-            let new_idx = EntityReferenceIdxModel {
-                entity_reference_id: entity_ref.id,
-                person_id: entity_ref.person_id,
-                reference_external_id_hash,
-                version: new_version,
-                hash: new_hash,
-            };
-            self.entity_reference_idx_cache
-                .read()
-                .await
-                .update(new_idx);
-        } else {
-            // INSERT
-            let version = 0;
-            let audit_model = EntityReferenceAuditModel {
-                entity_reference_id: entity_ref.id,
-                version,
-                hash: new_hash,
-                person_id: entity_ref.person_id,
-                entity_role: entity_ref.entity_role,
-                reference_external_id: entity_ref.reference_external_id.clone(),
-                reference_details_l1: entity_ref.reference_details_l1.clone(),
-                reference_details_l2: entity_ref.reference_details_l2.clone(),
-                reference_details_l3: entity_ref.reference_details_l3.clone(),
-                audit_log_id,
-            };
-
-            let query1 = sqlx::query(
-                r#"
-                INSERT INTO entity_reference_audit (entity_reference_id, version, hash, person_id, entity_role, reference_external_id, reference_details_l1, reference_details_l2, reference_details_l3, audit_log_id)
-                VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
-                "#,
-            )
-            .bind(audit_model.entity_reference_id)
-            .bind(audit_model.version)
-            .bind(audit_model.hash)
-            .bind(audit_model.person_id)
-            .bind(audit_model.entity_role)
-            .bind(audit_model.reference_external_id.as_str())
-            .bind(
-                audit_model
-                    .reference_details_l1
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(
-                audit_model
-                    .reference_details_l2
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(
-                audit_model
-                    .reference_details_l3
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(audit_model.audit_log_id);
-
-            let query2 = sqlx::query(
-                r#"
-                INSERT INTO entity_reference (id, person_id, entity_role, reference_external_id, reference_details_l1, reference_details_l2, reference_details_l3)
-                VALUES ($1, $2, $3::person_entity_type, $4, $5, $6, $7)
-                "#,
-            )
-            .bind(entity_ref.id)
-            .bind(entity_ref.person_id)
-            .bind(entity_ref.entity_role)
-            .bind(entity_ref.reference_external_id.as_str())
-            .bind(
-                entity_ref
-                    .reference_details_l1
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(
-                entity_ref
-                    .reference_details_l2
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            )
-            .bind(
-                entity_ref
-                    .reference_details_l3
-                    .as_ref()
-                    .map(|s| s.as_str()),
-            );
-
-            let query3 = sqlx::query(
-                r#"
-                INSERT INTO entity_reference_idx (entity_reference_id, person_id, version, hash)
-                VALUES ($1, $2, $3, $4)
-                "#,
-            )
-            .bind(entity_ref.id)
-            .bind(entity_ref.person_id)
-            .bind(version)
-            .bind(new_hash);
-
-            match &self.executor {
-                Executor::Pool(pool) => {
-                    query1.execute(&**pool).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                    query2.execute(&**pool).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                    query3.execute(&**pool).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                }
-                Executor::Tx(tx) => {
-                    let mut tx = tx.lock().await;
-                    query1.execute(&mut **tx).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                    query2.execute(&mut **tx).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                    query3.execute(&mut **tx).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?;
-                }
-            }
-
-            let new_idx = EntityReferenceIdxModel {
-                entity_reference_id: entity_ref.id,
-                person_id: entity_ref.person_id,
-                reference_external_id_hash,
-                version,
-                hash: new_hash,
-            };
-            self.entity_reference_idx_cache.read().await.add(new_idx);
-        }
-
-        Ok(entity_ref)
-    }
-
-    async fn load(&self, id: Uuid) -> EntityReferenceResult<EntityReferenceModel> {
-        let query = sqlx::query(
-            r#"
-            SELECT * FROM entity_reference WHERE id = $1
-            "#,
-        )
-        .bind(id);
-
-        let row = match &self.executor {
-            Executor::Pool(pool) => query.fetch_one(&**pool).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?,
-            Executor::Tx(tx) => {
-                let mut tx = tx.lock().await;
-                query.fetch_one(&mut **tx).await.map_err(|e| EntityReferenceRepositoryError::RepositoryError(e.into()))?
-            }
-        };
-
-        EntityReferenceModel::try_from_row(&row).map_err(EntityReferenceRepositoryError::RepositoryError)
-    }
-
-    async fn find_by_id(
-        &self,
-        id: Uuid,
-    ) -> EntityReferenceResult<Option<EntityReferenceIdxModel>> {
-        Ok(self
-            .entity_reference_idx_cache
-            .read()
-            .await
-            .get_by_primary(&id))
-    }
-
-    async fn find_by_person_id(
-        &self,
-        person_id: Uuid,
-        page: i32,
-        page_size: i32,
-    ) -> EntityReferenceResult<Vec<EntityReferenceIdxModel>> {
-        let cache = self.entity_reference_idx_cache.read().await;
-        if let Some(ids) = cache.get_by_person_id(&person_id) {
-            let start = ((page - 1) * page_size) as usize;
-            let end = (start + page_size as usize).min(ids.len());
-            if start >= ids.len() {
-                return Ok(Vec::new());
-            }
-            let mut refs = Vec::with_capacity(end - start);
-            for id in &ids[start..end] {
-                if let Some(model) = cache.get_by_primary(id) {
-                    refs.push(model);
-                }
-            }
-            Ok(refs)
-        } else {
-            Ok(Vec::new())
-        }
-    }
-
-    async fn find_by_reference_external_id(
-        &self,
-        reference_external_id: &str,
-        page: i32,
-        page_size: i32,
-    ) -> EntityReferenceResult<Vec<EntityReferenceIdxModel>> {
-        let mut hasher = XxHash64::with_seed(0);
-        hasher.write(reference_external_id.as_bytes());
-        let hash = hasher.finish() as i64;
-
-        let cache = self.entity_reference_idx_cache.read().await;
-        if let Some(ids) = cache.get_by_reference_external_id_hash(&hash) {
-            let start = ((page - 1) * page_size) as usize;
-            let end = (start + page_size as usize).min(ids.len());
-            if start >= ids.len() {
-                return Ok(Vec::new());
-            }
-            let mut refs = Vec::with_capacity(end - start);
-            for id in &ids[start..end] {
-                if let Some(model) = cache.get_by_primary(id) {
-                    refs.push(model);
-                }
-            }
-            Ok(refs)
-        } else {
-            Ok(Vec::new())
-        }
-    }
-
-    async fn find_by_ids(
-        &self,
-        ids: &[Uuid],
-    ) -> EntityReferenceResult<Vec<EntityReferenceIdxModel>> {
-        let cache = self.entity_reference_idx_cache.read().await;
-        let mut refs = Vec::with_capacity(ids.len());
-        for id in ids {
-            if let Some(model) = cache.get_by_primary(id) {
-                refs.push(model);
-            }
-        }
-        Ok(refs)
-    }
-
-    async fn exists_by_id(&self, id: Uuid) -> EntityReferenceResult<bool> {
-        Ok(self
-            .entity_reference_idx_cache
-            .read()
-            .await
-            .get_by_primary(&id)
-            .is_some())
-    }
-
-    async fn find_ids_by_person_id(
-        &self,
-        person_id: Uuid,
-    ) -> EntityReferenceResult<Vec<Uuid>> {
-        let cache = self.entity_reference_idx_cache.read().await;
-        if let Some(ids) = cache.get_by_person_id(&person_id) {
-            Ok(ids.clone())
-        } else {
-            Ok(Vec::new())
-        }
-    }
-
-    async fn exist_by_ids(
-        &self,
-        ids: &[Uuid],
-    ) -> EntityReferenceResult<Vec<(Uuid, bool)>> {
-        if ids.is_empty() {
-            return Ok(Vec::new());
-        }
-        let cache = self.entity_reference_idx_cache.read().await;
-        let mut result = Vec::with_capacity(ids.len());
-        for &id in ids {
-            result.push((id, cache.get_by_primary(&id).is_some()));
-        }
-        Ok(result)
-    }
-}
-
-#[async_trait]
-impl TransactionAware for EntityReferenceRepositoryImpl {
-    async fn on_commit(&self) -> BankingResult<()> {
-        self.entity_reference_idx_cache
-            .read()
-            .await
-            .on_commit()
-            .await
-    }
-
-    async fn on_rollback(&self) -> BankingResult<()> {
-        self.entity_reference_idx_cache
-            .read()
-            .await
-            .on_rollback()
-            .await
-    }
-}
-
-pub struct TransactionAwareEntityReferenceIdxModelCache {
-    shared_cache: Arc<RwLock<EntityReferenceIdxModelCache>>,
-    local_additions: RwLock<HashMap<Uuid, EntityReferenceIdxModel>>,
-    local_updates: RwLock<HashMap<Uuid, EntityReferenceIdxModel>>,
-    local_deletions: RwLock<HashSet<Uuid>>,
-}
-
-impl TransactionAwareEntityReferenceIdxModelCache {
-    pub fn new(shared_cache: Arc<RwLock<EntityReferenceIdxModelCache>>) -> Self {
-        Self {
-            shared_cache,
-            local_additions: RwLock::new(HashMap::new()),
-            local_updates: RwLock::new(HashMap::new()),
-            local_deletions: RwLock::new(HashSet::new()),
-        }
-    }
-
-    pub fn add(&self, item: EntityReferenceIdxModel) {
-        let primary_key = item.entity_reference_id;
-        self.local_deletions.write().remove(&primary_key);
-        self.local_additions.write().insert(primary_key, item);
-    }
-
-    pub fn update(&self, item: EntityReferenceIdxModel) {
-        let primary_key = item.entity_reference_id;
-        self.local_deletions.write().remove(&primary_key);
-        if let Some(local_item) = self.local_additions.write().get_mut(&primary_key) {
-            *local_item = item;
-            return;
-        }
-        self.local_updates.write().insert(primary_key, item);
-    }
-
-    pub fn remove(&self, primary_key: &Uuid) {
-        if self.local_additions.write().remove(primary_key).is_none() {
-            self.local_deletions.write().insert(*primary_key);
-        }
-        self.local_updates.write().remove(primary_key);
-    }
-
-    pub fn get_by_primary(&self, primary_key: &Uuid) -> Option<EntityReferenceIdxModel> {
-        if self.local_deletions.read().contains(primary_key) {
-            return None;
-        }
-        if let Some(item) = self.local_additions.read().get(primary_key) {
-            return Some(item.clone());
-        }
-        if let Some(item) = self.local_updates.read().get(primary_key) {
-            return Some(item.clone());
-        }
-        self.shared_cache.read().get_by_primary(primary_key)
-    }
-
-    pub fn get_by_person_id(&self, person_id: &Uuid) -> Option<Vec<Uuid>> {
-        let shared_cache = self.shared_cache.read();
-        let local_additions = self.local_additions.read();
-        let local_updates = self.local_updates.read();
-        let local_deletions = self.local_deletions.read();
-
-        let mut result_ids: HashSet<Uuid> = HashSet::new();
-
-        if let Some(ids) = shared_cache.get_by_person_id(person_id) {
-            for id in ids {
-                if !local_deletions.contains(id) {
-                    result_ids.insert(*id);
-                }
-            }
-        }
-
-        for updated_item in local_updates.values() {
-            if let Some(original_item) =
-                shared_cache.get_by_primary(&updated_item.entity_reference_id)
-            {
-                if original_item.person_id == *person_id && updated_item.person_id != *person_id {
-                    result_ids.remove(&updated_item.entity_reference_id);
-                }
-            }
-        }
-
-        for item in local_additions.values() {
-            if item.person_id == *person_id {
-                result_ids.insert(item.entity_reference_id);
-            }
-        }
-        for item in local_updates.values() {
-            if item.person_id == *person_id {
-                result_ids.insert(item.entity_reference_id);
-            }
-        }
-
-        if result_ids.is_empty() {
-            None
-        } else {
-            Some(result_ids.into_iter().collect())
-        }
-    }
-
-    pub fn get_by_reference_external_id_hash(&self, hash: &i64) -> Option<Vec<Uuid>> {
-        let shared_cache = self.shared_cache.read();
-        let local_additions = self.local_additions.read();
-        let local_updates = self.local_updates.read();
-        let local_deletions = self.local_deletions.read();
-
-        let mut result_ids: HashSet<Uuid> = HashSet::new();
-
-        if let Some(ids) = shared_cache.get_by_reference_external_id_hash(hash) {
-            for id in ids {
-                if !local_deletions.contains(id) {
-                    result_ids.insert(*id);
-                }
-            }
-        }
-
-        for updated_item in local_updates.values() {
-            if let Some(original_item) =
-                shared_cache.get_by_primary(&updated_item.entity_reference_id)
-            {
-                if original_item.reference_external_id_hash == *hash
-                    && updated_item.reference_external_id_hash != *hash
-                {
-                    result_ids.remove(&updated_item.entity_reference_id);
-                }
-            }
-        }
-
-        for item in local_additions.values() {
-            if item.reference_external_id_hash == *hash {
-                result_ids.insert(item.entity_reference_id);
-            }
-        }
-        for item in local_updates.values() {
-            if item.reference_external_id_hash == *hash {
-                result_ids.insert(item.entity_reference_id);
-            }
-        }
-
-        if result_ids.is_empty() {
-            None
-        } else {
-            Some(result_ids.into_iter().collect())
-        }
-    }
-}
-
-#[async_trait]
-impl TransactionAware for TransactionAwareEntityReferenceIdxModelCache {
-    async fn on_commit(&self) -> BankingResult<()> {
-        let mut shared_cache = self.shared_cache.write();
-        let mut local_additions = self.local_additions.write();
-        let mut local_updates = self.local_updates.write();
-        let mut local_deletions = self.local_deletions.write();
-
-        for item in local_additions.values() {
-            shared_cache.add(item.clone());
-        }
-        for item in local_updates.values() {
-            shared_cache.update(item.clone());
-        }
-        for primary_key in local_deletions.iter() {
-            shared_cache.remove(primary_key);
-        }
-
-        local_additions.clear();
-        local_updates.clear();
-        local_deletions.clear();
-        Ok(())
-    }
-
-    async fn on_rollback(&self) -> BankingResult<()> {
-        self.local_additions.write().clear();
-        self.local_updates.write().clear();
-        self.local_deletions.write().clear();
-        Ok(())
-    }
-}
-
-impl TryFromRow<PgRow> for EntityReferenceModel {
-    fn try_from_row(row: &PgRow) -> Result<Self, Box<dyn Error + Send + Sync>> {
-        Ok(EntityReferenceModel {
-            id: row.get("id"),
-            person_id: row.get("person_id"),
-            entity_role: row.get("entity_role"),
-            reference_external_id: get_heapless_string(row, "reference_external_id")?,
-            reference_details_l1: get_optional_heapless_string(
-                row,
-                "reference_details_l1",
-            )?,
-            reference_details_l2: get_optional_heapless_string(
-                row,
-                "reference_details_l2",
-            )?,
-            reference_details_l3: get_optional_heapless_string(
-                row,
-                "reference_details_l3",
-            )?,
-        })
-    }
-}
-
-impl TryFromRow<PgRow> for EntityReferenceIdxModel {
-    fn try_from_row(row: &PgRow) -> Result<Self, Box<dyn Error + Send + Sync>> {
-        Ok(EntityReferenceIdxModel {
-            entity_reference_id: row.get("entity_reference_id"),
-            person_id: row.get("person_id"),
-            reference_external_id_hash: row.get("reference_external_id_hash"),
-            version: row.get("version"),
-            hash: row.get("hash"),
-        })
-    }
-}
