+++ banking-db-postgres/src/repository/person/mod.rs
@@ -1,7 +1,6 @@
 pub mod country_repository;
 pub mod country_subdivision_repository;
-pub mod entity_reference_repository_batch_impl;
-pub mod entity_reference_repository_impl;
+pub mod entity_reference_repository;
 pub mod locality_repository_impl;
 pub mod locality_repository_batch_impl;
 pub mod location_repository_batch_impl;
