+++ banking-db-postgres/src/repository/unit_of_work_impl.rs
@@ -17,7 +17,7 @@ use crate::repository::{
     executor::Executor,
     person::country_repository::repo_impl::CountryRepositoryImpl,
     person::country_subdivision_repository::CountrySubdivisionRepositoryImpl,
-    person::entity_reference_repository_impl::EntityReferenceRepositoryImpl,
+    person::entity_reference_repository::EntityReferenceRepositoryImpl,
     person::locality_repository_impl::LocalityRepositoryImpl,
     person::location_repository_impl::LocationRepositoryImpl,
     person::person_repository_impl::PersonRepositoryImpl,
