+++ banking-db-postgres/tests/suites/person/mod.rs
@@ -1,7 +1,5 @@
-// pub mod entity_reference_repository_tests;
 // pub mod location_repository_tests;
 // pub mod locality_repository_tests;
-pub mod entity_reference_batch_operations_test;
 // pub mod locality_batch_operations_test;
 pub mod person_repository_tests;
 pub mod person_batch_operations_test;
